ChicagoTransitAPI
This is an example of how I can implement the use of external APIs into my development. This specific program is used to know the time of trains. This was also made in an earlier version of python so in my attempt to turn it into something worthy of the time of Python 3.6 it clearly had a stroke and killed itself. Enjoy anyways because im risking legal trouble by putting my api key in public view.