TitanAIO
This is my best attempt at creating a browser-task-based AIO bot. This is the only bot/automation I have ever used that has helped me cop shoes (that I have created).  This also uses the saving of tasks through unencrypted files so please for the love of god don't open program files for this when someone is looking because your nike accounts will be gone within the day.